src/requirements.txt
discord.py>=1.2.5
aiosqlite
